Switch to using BigDecimal

+++ app/model/ConversionResponse.scala
@@ -5,7 +5,7 @@ import play.api.libs.json.{Format, Json}
 /**
   * Response payload
   */
-case class ConversionResponse(exchange: Double, amount: Double, original: Double)
+case class ConversionResponse(exchange: BigDecimal, amount: BigDecimal, original: BigDecimal)
 
 object ConversionResponse {
   implicit val conversionResponseFormat: Format[ConversionResponse] = Json.format[ConversionResponse]
